backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-bcrypt==1.0.1
ultralytics==8.1.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.26.2
matplotlib==3.8.2
python-dotenv==1.0.0
